made it cleaner in separate file
