+++ src/components/union.tsx
@@ -21,6 +21,7 @@ import flatten from '@turf/flatten';
 import flattenEach from '@turf/meta';
 import Loading from './loading';
 import { modalStyle } from './styledComponents';
+import { flattenFeatures } from '../utils/flattenFeatures';
 
 function Union(props: {
   handleCloseModal: () => void;
@@ -55,39 +56,8 @@ function Union(props: {
       const layer1 = selectedLayer1.geoJSON;
       const layer2 = selectedLayer2.geoJSON;
 
-      const flattened1: FeatureCollection = {
-        type: 'FeatureCollection',
-        features: [],
-      };
+      const { flattened1, flattened2 } = flattenFeatures(layer1, layer2);
 
-      const flattened2: FeatureCollection = {
-        type: 'FeatureCollection',
-        features: [],
-      };
-
-      //Flatten if there are MultiPolygons(to make dissolve work)
-      layer1.features.forEach((feature) => {
-        if (feature.geometry.type === 'MultiPolygon') {
-          const tempGeom = flatten(feature.geometry);
-          tempGeom.features.forEach((poly) => {
-            flattened1.features.push(poly);
-            console.log('pushing feature:', poly);
-          });
-        } else {
-          flattened1.features.push(feature);
-        }
-      });
-      layer2.features.forEach((feature) => {
-        if (feature.geometry.type === 'MultiPolygon') {
-          const tempGeom = flatten(feature.geometry);
-          tempGeom.features.forEach((poly) => {
-            flattened2.features.push(poly);
-            console.log('pushing feature:', poly);
-          });
-        } else {
-          flattened2.features.push(feature);
-        }
-      });
       const dissolved1 = dissolve(flattened1 as FeatureCollection<Polygon, Properties>);
       const dissolved2 = dissolve(flattened2 as FeatureCollection<Polygon, Properties>);
